ajust code for sea level temporal mean
